+++ libs/partners/ollama/langchain_ollama/embeddings.py
@@ -12,7 +12,7 @@ from pydantic import (
 )
 from typing_extensions import Self
 
-from .utils import validate_model
+from ._utils import validate_model
 
 
 class OllamaEmbeddings(BaseModel, Embeddings):
